requirements.txt
beautifulsoup4==4.10.0
emoji==1.6.3
nltk==3.6.7
python_telegram_bot==13.10
requests==2.25.1
Wand==0.6.7
